requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
tensorflow==2.12.0
Pillow==9.5.0
numpy==1.24.3
gunicorn==20.1.0